+++ trade/detector.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from database.manager import Manager, transferring_row
 from database.models import UnsuitableLevelsDB
 from settings import LONG, SHORT
@@ -7,7 +9,7 @@ from trade.bot_request import Market
 class LevelDetector:
     """The class that checks for compliance with price requirements."""
     @staticmethod
-    def check_level(ticker: str, level: float, trend: str) -> bool:
+    def check_level(ticker: str, level: Decimal, trend: str) -> bool:
         """The method that checks the new levels for compliance when they are entered."""
         if trend == LONG:
             return level > Market.get_mark_price(ticker) and level not in Manager.get_level_by_trend(ticker, trend)
@@ -15,7 +17,7 @@ class LevelDetector:
             return level < Market.get_mark_price(ticker) and level not in Manager.get_level_by_trend(ticker, trend)
 
     @staticmethod
-    def check_levels(id: int, ticker: str, level: float, trend: str, **kwargs) -> None:
+    def check_levels(id: int, ticker: str, level: Decimal, trend: str, **kwargs) -> None:
         """
         A method that checks the levels which
         are already written to the database for compliance.
